requirements.txt
fastapi
uvicorn
whisperx
python-multipart
pydantic
numpy
ffmpeg-python